models/pyproject.toml
[project]
name = "models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tables>=3.10.2",
    "matplotlib>=3.10.8",
    "numpy>=2.2.0",
    "pandas>=2.3.3",
    "pyarrow>=20.0.0",
    "pyyaml>=6.0.2",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
]
